+++ src/types.ts
@@ -1,0 +1,11 @@
+const enum VSCODE_COMMANDS {
+  INIT_WORKSPACE = "grafana.initiateWorkspace",
+  DASHBOARD_DOWNLOAD = "grafana.downloadDashboardJSON",
+  DASHBOARD_UPLOAD = "grafana.uploadDashboardJSON",
+}
+
+const enum GLOBAL_STATE_VARIABLE {
+  URL = "grafana_url",
+  USERNAME = "grafana_username",
+  PASSWORD = "grafana_password",
+}